+++ delete_dataset.go
@@ -209,6 +209,10 @@ func (dd *DeleteDataset) ToSQL() (sql string, params []interface{}, err error) {
 // Appends this Dataset's DELETE statement to the SQLBuilder
 // This is used internally when using deletes in CTEs
 func (dd *DeleteDataset) AppendSQL(b sb.SQLBuilder) {
+	if dd.err != nil {
+		b.SetError(dd.err)
+		return
+	}
 	dd.dialect.ToDeleteSQL(b, dd.GetClauses())
 }