+++ HISTORY.md
@@ -1,6 +1,7 @@
 # v9.5.1
 
 * [FIXED] Unable to execute union with order by expression [#185](https://github.com/doug-martin/goqu/issues/185)
+* [FIXED] SelectDataset From with Error [#183](https://github.com/doug-martin/goqu/issues/183)
 
 # v9.5.0